Add search user endpoint. Refactor user controllers into multiple files. Rename pokerboard route. Update findUsers function.
