+++ src/controllers/user.controller.registration.ts
@@ -1,52 +1,15 @@
 import { StatusCodes } from 'http-status-codes';
 import { Request, Response } from 'express';
 import * as jwt from 'jsonwebtoken';
-import { Not } from 'typeorm';
 
 import { EmailSubject } from '../constants/enums';
-import {
-  EmailMessages,
-  ErrorMessages,
-  ResponseMessages,
-  ValidationMessages,
-} from '../constants/message';
-import { createToken, findAndUpdateToken } from '../entity/token/repository';
+import { EmailMessages, ErrorMessages, ResponseMessages } from '../constants/message';
 import { User } from '../entity/user/model';
-import { createUser, findAndUpdateUser, findUser, saveUser } from '../entity/user/repository';
+import { createUser, findAndUpdateUser } from '../entity/user/repository';
 import config from '../settings/config';
-import { comparePassword, hashPassword } from '../utils/auth';
+import { hashPassword } from '../utils/auth';
 import { makeResponse } from '../utils/common';
 import { sendMail } from '../utils/notification';
-
-export const getUser = async (req: Request, res: Response) => {
-  const { password, ...user } = req.user; // eslint-disable-line @typescript-eslint/no-unused-vars
-  if (!user.isVerified) {
-    const errorData = { verify: ErrorMessages.ACCOUNT_NOT_VERIFIED };
-    return res
-      .status(StatusCodes.BAD_REQUEST)
-      .json(makeResponse(false, ErrorMessages.ACCOUNT_NOT_VERIFIED, errorData));
-  } else {
-    return res
-      .status(StatusCodes.OK)
-      .json(makeResponse(true, ResponseMessages.GET_USER_SUCCESS, user));
-  }
-};
-
-export const loginUser = async (req: Request, res: Response) => {
-  const user = req.user;
-  const { password, ...userData } = user;
-
-  try {
-    await comparePassword(req.body.password, password);
-  } catch (error) {
-    return res.status(error.statusCode).json(makeResponse(false, error.message, error.data));
-  }
-
-  const { token } = await createToken(user);
-  res
-    .status(StatusCodes.OK)
-    .json(makeResponse(true, ResponseMessages.LOGIN_SUCCESS, { token, userData }));
-};
 
 export const registerUser = async (req: Request, res: Response) => {
   const body = req.body;
@@ -70,12 +33,6 @@ export const registerUser = async (req: Request, res: Response) => {
   }
 };
 
-export const logoutUser = async (req: Request, res: Response) => {
-  const currentDate = new Date();
-  await findAndUpdateToken({ id: req.token.id }, { expiredAt: currentDate, isActive: false });
-  return res.status(StatusCodes.OK).json(makeResponse(true, ResponseMessages.LOGOUT_SUCCESS));
-};
-
 export const reverifyUser = async (req: Request, res: Response) => {
   const email = req.body.email;
   const token = jwt.sign({ email: email, id: req.user.id }, config.JWT.SECRECT, {
@@ -85,37 +42,6 @@ export const reverifyUser = async (req: Request, res: Response) => {
   return res
     .status(StatusCodes.OK)
     .json(makeResponse(true, ResponseMessages.VERIFICATION_LINK_SENT));
-};
-
-export const updateUser = async (req: Request, res: Response) => {
-  const { firstName, lastName, username } = req.body;
-  const user = req.user;
-
-  if (firstName) {
-    user.firstName = firstName;
-  }
-
-  if (lastName) {
-    user.lastName = lastName;
-  }
-
-  if (username) {
-    const userWithSameUsername = await findUser({ username: username, id: Not(user.id) });
-    if (userWithSameUsername) {
-      const errorData = { username: ValidationMessages.USERNAME_ALREADY_EXISTS };
-      return res
-        .status(StatusCodes.BAD_REQUEST)
-        .json(makeResponse(false, ValidationMessages.USERNAME_ALREADY_EXISTS, errorData));
-    } else {
-      user.username = username;
-    }
-  }
-
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  const { password, ...userData } = await saveUser(user);
-  return res
-    .status(StatusCodes.OK)
-    .json(makeResponse(true, ResponseMessages.UPDATE_USER_SUCCESS, userData));
 };
 
 export const verifyUser = async (req: Request, res: Response) => {
